Feature/configure logger (#9)
* Add static functions to switch verbosity levels

Just a simple UI for changing to fatal (which will literally print nothing) or low (which is the quite verbose default of xbot/opensot)

* Set fatal to default

So nothing have to be changed w.r.t. to configuring the logger. Set the verbosity to low with the added static functions to recreate "old" behavior on runtime